requirements.txt
requests~=2.31.0
torch~=2.0.1+cu118
Pillow~=9.5.0
quart
quart-cors
openai